refactor: move Tasklist related indices to index package
