+++ tasklist/els-schema/src/main/java/io/camunda/tasklist/store/elasticsearch/TaskMetricsStoreElasticSearch.java
@@ -8,16 +8,16 @@
 package io.camunda.tasklist.store.elasticsearch;
 
 import static io.camunda.tasklist.util.ElasticsearchUtil.LENIENT_EXPAND_OPEN_IGNORE_THROTTLED;
-import static io.camunda.webapps.schema.descriptors.tasklist.index.TasklistMetricIndex.EVENT;
-import static io.camunda.webapps.schema.descriptors.tasklist.index.TasklistMetricIndex.EVENT_TIME;
-import static io.camunda.webapps.schema.descriptors.tasklist.index.TasklistMetricIndex.VALUE;
+import static io.camunda.webapps.schema.descriptors.index.TasklistMetricIndex.EVENT;
+import static io.camunda.webapps.schema.descriptors.index.TasklistMetricIndex.EVENT_TIME;
+import static io.camunda.webapps.schema.descriptors.index.TasklistMetricIndex.VALUE;
 import static org.elasticsearch.index.query.QueryBuilders.boolQuery;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.camunda.tasklist.data.conditionals.ElasticSearchCondition;
 import io.camunda.tasklist.exceptions.TasklistRuntimeException;
 import io.camunda.tasklist.store.TaskMetricsStore;
-import io.camunda.webapps.schema.descriptors.tasklist.index.TasklistMetricIndex;
+import io.camunda.webapps.schema.descriptors.index.TasklistMetricIndex;
 import io.camunda.webapps.schema.entities.MetricEntity;
 import io.camunda.webapps.schema.entities.usertask.TaskEntity;
 import java.io.IOException;
@@ -64,7 +64,7 @@ public class TaskMetricsStoreElasticSearch implements TaskMetricsStore {
   private ObjectMapper objectMapper;
 
   @Override
-  public void registerTaskCompleteEvent(TaskEntity task) {
+  public void registerTaskCompleteEvent(final TaskEntity task) {
     final MetricEntity metric = createTaskCompleteEvent(task);
     final boolean inserted = insert(metric);
     if (!inserted) {
@@ -74,24 +74,9 @@ public class TaskMetricsStoreElasticSearch implements TaskMetricsStore {
     }
   }
 
-  private boolean insert(MetricEntity entity) {
-    try {
-      final IndexRequest request =
-          new IndexRequest(index.getFullQualifiedName())
-              .id(entity.getId())
-              .source(objectMapper.writeValueAsString(entity), XContentType.JSON);
-
-      final IndexResponse response = esClient.index(request, RequestOptions.DEFAULT);
-      return response.status() == RestStatus.CREATED;
-    } catch (IOException e) {
-      LOGGER.error(e.getMessage(), e);
-      throw new TasklistRuntimeException("Error while trying to upsert entity: " + entity);
-    }
-  }
-
   @Override
   public List<String> retrieveDistinctAssigneesBetweenDates(
-      OffsetDateTime startTime, OffsetDateTime endTime) {
+      final OffsetDateTime startTime, final OffsetDateTime endTime) {
 
     final BoolQueryBuilder rangeQuery =
         boolQuery()
@@ -115,7 +100,7 @@ public class TaskMetricsStoreElasticSearch implements TaskMetricsStore {
       }
 
       final Aggregation group = aggregations.get(ASSIGNEE);
-      if (!(group instanceof ParsedStringTerms terms)) {
+      if (!(group instanceof final ParsedStringTerms terms)) {
         throw new TasklistRuntimeException("Unexpected response for aggregations");
       }
 
@@ -123,14 +108,29 @@ public class TaskMetricsStoreElasticSearch implements TaskMetricsStore {
           (List<ParsedStringTerms.ParsedBucket>) terms.getBuckets();
 
       return buckets.stream().map(it -> String.valueOf(it.getKey())).collect(Collectors.toList());
-    } catch (IOException e) {
+    } catch (final IOException e) {
       LOGGER.error("Error while retrieving assigned users between dates from index: " + index, e);
       final String message = "Error while retrieving assigned users between dates";
       throw new TasklistRuntimeException(message);
     }
   }
 
-  private MetricEntity createTaskCompleteEvent(TaskEntity task) {
+  private boolean insert(final MetricEntity entity) {
+    try {
+      final IndexRequest request =
+          new IndexRequest(index.getFullQualifiedName())
+              .id(entity.getId())
+              .source(objectMapper.writeValueAsString(entity), XContentType.JSON);
+
+      final IndexResponse response = esClient.index(request, RequestOptions.DEFAULT);
+      return response.status() == RestStatus.CREATED;
+    } catch (final IOException e) {
+      LOGGER.error(e.getMessage(), e);
+      throw new TasklistRuntimeException("Error while trying to upsert entity: " + entity);
+    }
+  }
+
+  private MetricEntity createTaskCompleteEvent(final TaskEntity task) {
     return new MetricEntity()
         .setEvent(EVENT_TASK_COMPLETED_BY_ASSIGNEE)
         .setValue(task.getAssignee())
