+++ tasklist/importer-870/src/main/java/io/camunda/tasklist/zeebeimport/v870/processors/os/FormZeebeRecordProcessorOpenSearch.java
@@ -12,7 +12,7 @@ import io.camunda.tasklist.exceptions.PersistenceException;
 import io.camunda.tasklist.util.ConversionUtils;
 import io.camunda.tasklist.util.OpenSearchUtil;
 import io.camunda.tasklist.zeebeimport.v870.record.value.deployment.FormRecordImpl;
-import io.camunda.webapps.schema.descriptors.tasklist.index.FormIndex;
+import io.camunda.webapps.schema.descriptors.index.FormIndex;
 import io.camunda.webapps.schema.entities.form.FormEntity;
 import io.camunda.zeebe.protocol.record.Record;
 import io.camunda.zeebe.protocol.record.intent.FormIntent;
