+++ tasklist/importer-870/src/main/java/io/camunda/tasklist/zeebeimport/v870/processors/os/ProcessZeebeRecordProcessorOpenSearch.java
@@ -14,8 +14,8 @@ import io.camunda.tasklist.util.ConversionUtils;
 import io.camunda.tasklist.zeebeimport.common.ProcessDefinitionDeletionProcessor;
 import io.camunda.tasklist.zeebeimport.util.XMLUtil;
 import io.camunda.tasklist.zeebeimport.v870.record.value.deployment.DeployedProcessImpl;
+import io.camunda.webapps.schema.descriptors.index.FormIndex;
 import io.camunda.webapps.schema.descriptors.operate.index.ProcessIndex;
-import io.camunda.webapps.schema.descriptors.tasklist.index.FormIndex;
 import io.camunda.webapps.schema.entities.ProcessEntity;
 import io.camunda.webapps.schema.entities.form.FormEntity;
 import io.camunda.zeebe.protocol.record.Record;
