+++ tasklist/importer-common/src/main/java/io/camunda/tasklist/zeebeimport/common/ProcessDefinitionDeletionProcessor.java
@@ -11,8 +11,8 @@ import io.camunda.tasklist.store.DraftVariableStore;
 import io.camunda.tasklist.store.FormStore;
 import io.camunda.tasklist.store.TaskStore;
 import io.camunda.tasklist.store.VariableStore;
+import io.camunda.webapps.schema.descriptors.index.FormIndex;
 import io.camunda.webapps.schema.descriptors.operate.index.ProcessIndex;
-import io.camunda.webapps.schema.descriptors.tasklist.index.FormIndex;
 import io.camunda.webapps.schema.descriptors.tasklist.template.DraftTaskVariableTemplate;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -58,8 +58,8 @@ public class ProcessDefinitionDeletionProcessor {
   @Autowired private DraftVariableStore draftVariableStore;
 
   public <T> List<T> createProcessDefinitionDeleteRequests(
-      String processDefinitionId,
-      BiFunction<String, String, T>
+      final String processDefinitionId,
+      final BiFunction<String, String, T>
           deleteRequestBuilder // (indexName, documentId) to DeleteRequest mapper
       ) {
     final Map<String, String> taskIdsToIndex =
@@ -99,14 +99,17 @@ public class ProcessDefinitionDeletionProcessor {
   }
 
   private <T> List<T> createDeleteRequestList(
-      List<String> ids, String indexName, BiFunction<String, String, T> deleteRequestBuilder) {
+      final List<String> ids,
+      final String indexName,
+      final BiFunction<String, String, T> deleteRequestBuilder) {
     return ids.stream()
         .map(id -> deleteRequestBuilder.apply(indexName, id))
         .collect(Collectors.toList());
   }
 
   private <T> List<T> createDeleteRequestList(
-      Map<String, String> idsToIndex, BiFunction<String, String, T> deleteRequestBuilder) {
+      final Map<String, String> idsToIndex,
+      final BiFunction<String, String, T> deleteRequestBuilder) {
     return idsToIndex.entrySet().stream()
         .map(entry -> deleteRequestBuilder.apply(entry.getValue(), entry.getKey()))
         .collect(Collectors.toList());
