+++ tasklist/importer-common/src/test/java/io/camunda/tasklist/zeebeimport/common/ProcessDefinitionDeletionProcessorTest.java
@@ -14,8 +14,8 @@ import io.camunda.tasklist.store.DraftVariableStore;
 import io.camunda.tasklist.store.FormStore;
 import io.camunda.tasklist.store.TaskStore;
 import io.camunda.tasklist.store.VariableStore;
+import io.camunda.webapps.schema.descriptors.index.FormIndex;
 import io.camunda.webapps.schema.descriptors.operate.index.ProcessIndex;
-import io.camunda.webapps.schema.descriptors.tasklist.index.FormIndex;
 import io.camunda.webapps.schema.descriptors.tasklist.template.DraftTaskVariableTemplate;
 import java.util.LinkedHashMap;
 import java.util.List;
