+++ zeebe/exporters/camunda-exporter/src/main/java/io/camunda/exporter/cache/form/OpenSearchFormCacheLoader.java
@@ -8,7 +8,7 @@
 package io.camunda.exporter.cache.form;
 
 import com.github.benmanes.caffeine.cache.CacheLoader;
-import io.camunda.webapps.schema.descriptors.tasklist.index.FormIndex;
+import io.camunda.webapps.schema.descriptors.index.FormIndex;
 import io.camunda.webapps.schema.entities.form.FormEntity;
 import java.io.IOException;
 import org.opensearch.client.opensearch.OpenSearchClient;
